+++ chartlets.js/packages/lib/src/plugins/mui/Tabs.tsx
@@ -9,10 +9,11 @@ import { Box } from "@/plugins/mui/Box";
 import { isString } from "@/utils/isString";
 import { isComponentState } from "@/types/state/component";
 
-interface TabState {
+interface TabState extends ComponentState {
   type: "Tab";
   label?: string;
   icon?: string;
+  iconPosition?: "bottom" | "end" | "start" | "top" | undefined;
   disabled?: boolean;
   children?: ComponentProps[];
 }
@@ -44,9 +45,9 @@ export function Tabs({
     }
   };
   return (
-    <MuiBox sx={{ width: "100%" }}>
+    <MuiBox sx={{ width: "100%" }} style={style}>
       <MuiBox sx={{ borderBottom: 1, borderColor: "divider" }}>
-        <MuiTabs id={id} style={style} value={value} onChange={handleChange}>
+        <MuiTabs id={id} value={value} onChange={handleChange}>
           {tabItems?.map((tab, index) => {
             const tabState = isComponentState(tab)
               ? (tab as TabState)
@@ -54,11 +55,13 @@ export function Tabs({
             return (
               <MuiTab
                 key={index}
+                style={tabState?.style}
                 label={tabState ? tabState.label : isString(tab) ? tab : ""}
                 icon={
                   tabState &&
                   tabState.icon && <MuiIcon>{tabState.icon}</MuiIcon>
                 }
+                iconPosition={tabState?.iconPosition}
                 disabled={disabled || (tabState && tabState.disabled)}
               />
             );
@@ -70,10 +73,11 @@ export function Tabs({
         return (
           value === index && (
             <Box
+              style={tabState?.style}
               key={index}
               type={type}
               onChange={onChange}
-              children={tabState?.children ?? undefined}
+              children={tabState?.children}
             />
           )
         );